moved db models away from db related functions
