+++ slg_desktop/__init__.py.orig
@@ -1,0 +1,1 @@
+__version__ = '0.1.0'
